requirements.txt
streamlit
networkx
matplotlib
arabic-reshaper
python-bidi
